chore(backend-types): rename handler.types and place in types folder
