+++ packages/backend/src/handlers/playground.ts
@@ -1,4 +1,4 @@
-import {GqlHandlerOptions} from './handler.types';
+import {GqlHandlerOptions} from '../types/handler';
 
 const html = (baseEndpoint: string) => `
 <!DOCTYPE html>
