+++ packages/backend/src/types/handler.ts
@@ -2,7 +2,7 @@ export interface GqlHandlerOptions {
     baseEndpoint: string;
     playgroundEndpoint: string;
     forwardUnmatchedRequestsToOrigin: boolean;
-    debug: boolean;
+    allowDebug: boolean;
     cors: boolean | object; // TODO: type this
     kvCache: boolean;
 }
